Improved ticketservice implementation and added tests

+++ src/main/java/uk/gov/dwp/uc/account/AccountService.java
@@ -1,7 +1,7 @@
 package uk.gov.dwp.uc.account;
 
 public interface AccountService {
-    boolean validateAccount(Long accountId);
-    long  checkBalance(Long accountId);
+    boolean validateAccount(long accountId);
+    int  checkBalance(long accountId);
 
 }
